+++ documentation/docs/src/lib/components/Form/DatePicker/DatePicker.stories.svelte
@@ -163,6 +163,8 @@ let value = $state<DateValue>(new CalendarDate(2026, 4, 7));
   args={{ triggerPlacement: 'before', popupProps: { align: 'start' } }}
 />
 
+<Story name="No Trigger Button" args={{ showTriggerButton: false }} />
+
 <Story name="Custom Trigger Icon" asChild>
   <DatePicker>
     {#snippet customTriggerIcon()}